refactor(export): domain 不再依賴 presentation 的 PDF builder
PdfEntryData / PdfLabels 本是純資料，移回 export/domain/models；builder 改
成透過 PdfDocumentBuilderFn typedef 注入（與既有 cover/fonts/share 全注入
的設計一致），trip_pdf_export_service 不再直接 new presentation 的
TripPdfDocumentBuilder。pipeline 於 presentation 層補上 buildDocument wiring，
providers 改由 domain re-export PdfLabels。
