+++ frontend/lib/features/export/presentation/default_pdf_export_pipeline.dart
@@ -5,6 +5,7 @@ import 'package:context_app/features/export/domain/models/pdf_export_result.dart
 import 'package:context_app/features/export/domain/services/place_image_downloader.dart';
 import 'package:context_app/features/export/domain/services/trip_pdf_export_service.dart';
 import 'package:context_app/features/export/presentation/cover_renderer.dart';
+import 'package:context_app/features/export/presentation/pdf_builder/trip_pdf_document_builder.dart';
 import 'package:context_app/features/trip/providers.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -41,6 +42,18 @@ Future<PdfExportResult> exportTripAsPdf({
       regular: await PdfGoogleFonts.notoSansTCRegular(),
       bold: await PdfGoogleFonts.notoSansTCBold(),
     ),
+    buildDocument:
+        ({
+          required trip,
+          required entries,
+          required coverPngBytes,
+          required fonts,
+          required labels,
+        }) => TripPdfDocumentBuilder(
+          regularFont: fonts.regular,
+          boldFont: fonts.bold,
+          labels: labels,
+        ).build(trip: trip, entries: entries, coverPngBytes: coverPngBytes),
     writeToTemp: _writeToTemp,
     share: _sharePdf,
   );
