+++ frontend/lib/features/export/presentation/pdf_builder/trip_pdf_document_builder.dart
@@ -1,41 +1,10 @@
 import 'dart:typed_data';
 
+import 'package:context_app/features/export/domain/models/pdf_entry_data.dart';
+import 'package:context_app/features/export/domain/models/pdf_labels.dart';
 import 'package:context_app/features/trip/domain/models/trip.dart';
 import 'package:pdf/pdf.dart';
 import 'package:pdf/widgets.dart' as pw;
-
-/// Per-entry data prepared for PDF rendering.
-///
-/// Works for both narration entries (with an optional story hook chip) and
-/// quick-guide entries (chips list is empty, address is null).
-class PdfEntryData {
-  final String title;
-  final String? address;
-  final DateTime date;
-  final String bodyText;
-  final List<String> chipLabels;
-  final Uint8List? imageBytes;
-
-  const PdfEntryData({
-    required this.title,
-    required this.date,
-    required this.bodyText,
-    this.address,
-    this.chipLabels = const [],
-    this.imageBytes,
-  });
-}
-
-/// Pre-translated strings used while building the PDF.
-class PdfLabels {
-  final String pageOfTotal;
-
-  const PdfLabels({required this.pageOfTotal});
-
-  String renderPageOfTotal(int index, int total) => pageOfTotal
-      .replaceAll('{index}', '$index')
-      .replaceAll('{total}', '$total');
-}
 
 /// Builds the binary PDF document for a Trip export.
 ///
