+++ frontend/lib/features/export/providers.dart
@@ -1,5 +1,5 @@
 // Trip PDF export feature 的公開介面。
 //
 // 依 CLAUDE.md 依賴規則，其他 feature 一律經由本檔引用 export 功能。
+export 'domain/models/pdf_labels.dart';
 export 'presentation/default_pdf_export_pipeline.dart';
-export 'presentation/pdf_builder/trip_pdf_document_builder.dart';
